Embed a popover into an html element

This is useful to add a more-verbose explanation to an element.

bs_embed_popover(tag, title = NULL, content = NULL, placement = "top",
  ...)

use_bs_popover()

Arguments

tag

htmltools::tag, button or link into which to embed a popover

title

character, title for the popover

content

character, content for the popover body

placement

character, placement of the popover with respect to the tag

...

other named arguments, passed to bs_set_data()

Value

htmltools::tag, modified copy of tag

Details

To activate the use of popovers in your page, you will need to call the use_bs_popover() function somewhere.

The verb "embed" is used to signify that we are embedding information into the tag. This implies that we can embed, at most, one "thing" into a particular tag. We should not, for example, embed both a tooltip and a popover into a tag.

See also

bs_embed_tooltip, http://getbootstrap.com/javascript/#popovers

Examples

library("htmltools")

tags$button(type = "button", class = "btn btn-default", "A button") %>%
  bs_embed_popover(title = "I'm a popover", content = "Really!")

#> <button type="button" class="btn btn-default" title="I&#39;m a popover" data-toggle="popover" data-content="Really!" data-placement="top">A button</button>

tags$button(type = "button", class = "btn btn-default", "Another button") %>%
  bs_embed_popover(
    title  = "Verbose popover",
    content = render_html_fragment(
      system.file("markdown", "modal.md", package = "bsplus")
    )
  )

#> /Applications/RStudio.app/Contents/MacOS/pandoc/pandoc +RTS -K512m -RTS modal.utf8.md --to html --from markdown+autolink_bare_uris+ascii_identifiers+tex_math_single_backslash --output /var/folders/rg/9kvy_lsn43z90d439n02nfw80000gn/T//RtmpEzVGge/file4d94624026f1 --smart --email-obfuscation none --self-contained --mathjax --standalone --section-divs --template /Library/Frameworks/R.framework/Versions/3.3/Resources/library/rmarkdown/rmd/fragment/default.html --no-highlight 

#> 
#> Output created: /var/folders/rg/9kvy_lsn43z90d439n02nfw80000gn/T//RtmpEzVGge/file4d94624026f1

#> <button type="button" class="btn btn-default" title="Verbose popover" data-toggle="popover" data-content="&lt;div id=&quot;th-level-headers-work-well&quot; class=&quot;section level4&quot;&gt; &lt;h4&gt;4th-level headers work well&lt;/h4&gt; &lt;p&gt;We can do our regular Markdown stuff in modal windows.&lt;/p&gt; &lt;ul&gt; &lt;li&gt;&lt;span class=&quot;math inline&quot;&gt;\(F = ma\)&lt;/span&gt;&lt;/li&gt; &lt;li&gt;&lt;span class=&quot;math inline&quot;&gt;\(E = mc^2\)&lt;/span&gt;&lt;/li&gt; &lt;/ul&gt; &lt;/div&gt;" data-placement="top">Another button</button>