%>% {bsplus}	R Documentation

Pipe

Description

Like dplyr, bsplus also uses the pipe function, %>% to turn function composition into a series of imperative statements.

Arguments

lhs, rhs
An object and a function to apply to it

[Package bsplus version 0.1.1 Index]