shinyInput_label_embed {bsplus}	R Documentation

Embed an element into the label of a Shiny-input tag

Description

The element embedded into the Shiny input will be pulled to the right edge of the label.

Usage

shinyInput_label_embed(tag, element)

Arguments

tag
Shiny input, such as shiny::numericInput

element
htmltools::tag to be embedded into label of tag

Details

To promote consistency, the following convention is proposed:

For links (activated by clicking), embed a shiny::icon("info-circle"); this is the default for shiny_iconlink. For elements activated by hovering, embed a shiny::icon("info").

Value

Shiny input, modified copy of tag

See Also

shiny_iconlink

Examples

library("shiny")

numericInput(inputId = "foo", label = "Enter a number", value = 0) %>%
  shinyInput_label_embed(
    shiny_iconlink() %>%
    bs_embed_popover(title = "Number", content = "Not a complex number")
  )

[Package bsplus version 0.1.1 Index]