shiny_iconlink {bsplus}	R Documentation

Create link containing Shiny icon

Description

You can use this helper function to wrap link element around a shiny::icon. It may be useful to attach a modal window to (or embed a popover into) into such a link.

Usage

shiny_iconlink(name = "info-circle", ...)

Arguments

name
character, name of the icon, passed to shiny::icon

...
other arguments passed to shiny::icon

Value

htmltools::tag, <a/>

See Also

shinyInput_label_embed, shiny::icon, bs_attach_modal, bs_embed_popover, bs_embed_tooltip

Examples

shiny_iconlink()

shiny_iconlink() %>%
  bs_embed_popover(title = "Help!", content = "I need somebody")

[Package bsplus version 0.1.1 Index]