Builds an accordion panel-group.

An accordion is a group of panels where, at most, one panel body is visible.

bs_accordion(id)

# S3 method for bsplus_accordion
bs_append(tag, title, content, ...)

# S3 method for bsplus_accordion
bs_set_opts(tag, panel_type = "primary",
  use_heading_link = TRUE, ...)

Arguments

id

character, id for div

tag

shiny.tag, accordion element

title

HTML or text, used for the panel heading

content

HTML or tagList, used for the panel body.

...

other args

panel_type

character, one of the standard Bootstrap types c("default", "primary", "success", "info", "warning", "danger")

use_heading_link

logical, indicates whether to use the entire panel heading as a link. If FALSE, uses construction shown in http://getbootstrap.com/javascript/#collapse-example-accordion; if TRUE, makes the entire panel heading clickable.

Value

bsplus_accordion object

Details

This constructor takes a single argument, an id. It is your responsibility to ensure that id is unique among html elements in your page. If you have non-unique id's, strange things may happen as your page is rendered.

The constructor is a part of a family of function to set the options for panels, and a function to append a panel to the panel group.

All of these functions return a bsplus_accordion object, so you can compose an accordion by piping. The bsplus_accordion class inherits from htmltools' shiny.tag class.

The bs_set_opts method is used to modify panels that have yet to be appended. There are two options that can be set for accordions:

panel_type

One of c("default", "primary", "success", "info", "warning", "danger"). These are the standard Bootstrap types; the default is "default".

use_heading_link

Indicates if the entire panel-heading shall be made clickable; the default is FALSE.

The append method is used to build and append a panel to an accordion. Such a panel is built using two additional arguments:

title

Usually text for the panel heading, but HTML can be used as well

content

Content for the panel body, can be HTML or an htmltools tagList

See also

http://getbootstrap.com/javascript/#collapse-example-accordion

Examples

bs_accordion(id = "meet_the_beatles") %>%
  bs_set_opts(panel_type = "success", use_heading_link = TRUE) %>%
  bs_append(title = "John Lennon", content = "Rhythm guitar, vocals") %>%
  bs_set_opts(panel_type = "info") %>%
  bs_append(title = "Paul McCartney", content = "Bass guitar, vocals")

#> <div id="meet_the_beatles" class="panel-group" role="tablist" aria-multiselectable="true">
#>   <div class="panel panel-success" id="meet_the_beatles-0">
#>     <div id="meet_the_beatles-0-heading" class="panel-heading" role="tab" data-toggle="collapse" data-target="#meet_the_beatles-0-collapse" data-parent="#meet_the_beatles" aria-expanded="true" aria-controls="meet_the_beatles-0-collapse" style="cursor: pointer;">
#>       <h4 class="panel-title">John Lennon</h4>
#>     </div>
#>     <div aria-labelledby="meet_the_beatles-0-heading" class="panel-collapse collapse in" id="meet_the_beatles-0-collapse" role="tabpanel">
#>       <div class="panel-body" style="">Rhythm guitar, vocals</div>
#>     </div>
#>   </div>
#>   <div class="panel panel-info" id="meet_the_beatles-1">
#>     <div id="meet_the_beatles-1-heading" class="panel-heading" role="tab" data-toggle="collapse" data-target="#meet_the_beatles-1-collapse" data-parent="#meet_the_beatles" aria-expanded="true" aria-controls="meet_the_beatles-1-collapse" style="cursor: pointer;">
#>       <h4 class="panel-title">Paul McCartney</h4>
#>     </div>
#>     <div id="meet_the_beatles-1-collapse" class="panel-collapse collapse" role="tabpanel" aria-labelledby="meet_the_beatles-1-heading">
#>       <div class="panel-body" style="">Bass guitar, vocals</div>
#>     </div>
#>   </div>
#> </div>