framework for accordion-sidebar

framework for accordion-sidebar

bs_accordion_sidebar(id, spec_side = c(width = 4, offset = 0),
  spec_main = c(width = 8, offset = 0), position = c("left", "right"))

# S3 method for bsplus_accordion_sidebar
bs_append(tag, title_side, content_side,
  content_main, ...)

# S3 method for bsplus_accordion_sidebar
bs_set_opts(tag,
  panel_type_active = "success", panel_type_inactive = "primary", ...)

Arguments

id

character, id for div and root it for included elements

spec_side

numeric, column specification for sidebar panels

spec_main

numeric, column specification for main panels

position

character, indicates where to put the sidebar panels with repspect to the main panels

tag

shiny.tag, accordion-sidebar element

title_side

character or HTML, stuff to put into sidebar title

content_side

character or HTML, stuff to put into sidebar content

content_main

character or HTML, stuff to put into main content

...

other args

panel_type_active

character, indicated bootstrap type for active-panel header

panel_type_inactive

character, indicated bootstrap type for inactive-panel header

Value

bsplus_accsidebar object

Examples

bs_accordion_sidebar(id = "meet_the_beatles")

#> <div class="row" id="meet_the_beatles">
#>   <div class="col-sm-4" id="meet_the_beatles-side">
#>     <div id="meet_the_beatles-side-accordion" class="panel-group" role="tablist" aria-multiselectable="true"></div>
#>   </div>
#>   <div class="col-sm-8" id="meet_the_beatles-main"></div>
#> </div>