Create and attach a collapsible div

This is useful for content that you may wish to be hidden when the page is initialized, but that can be revealed (and subsequently hidden) by clicking a button or a link.

bs_collapse(id, content = NULL, show = FALSE)

bs_attach_collapse(tag, id_collapse)

Arguments

id

character, unique id for modal window to create

content

character (HTML) or htmltools::tagList, content to be put into the div

show

logical, indicates if element is shown initially

tag

htmltools::tag, button or link to which to attach collapsible div

id_collapse

character, id of collapsible div to attach

Value

bs_collapse()

htmltools::tag, div

bs_attach_collapse()

htmltools::tag, modified copy of tag

Details

There are two parts to the system:

A collapsible div, created using bs_collapse()

At least one button or link to which the id of the collabsible div is attached, using bs_attach_collapse()

Keep in mind that you can attach a collapsible div to more than one button or link.

It is your responsibility to ensure that id is unique among html elements in your page. If you have non-unique id's, strange things may happen to your page.

See also

https://getbootstrap.com/javascript/#collapse

Examples

library("htmltools")

bs_collapse(id = "id_yeah", tags$p("Yeah Yeah Yeah"))

#> <div class="collapse" id="id_yeah">
#>   <p>Yeah Yeah Yeah</p>
#> </div>

tags$button("She Loves You") %>%
  bs_attach_collapse("id_yeah")

#> <button data-toggle="collapse" data-target="#id_yeah">She Loves You</button>