bs_embed_popover {bsplus}	R Documentation

Popover

Description

A popover can be a useful way to add a somewhat-verbose explanation to a tag.

Usage

bs_embed_popover(tag, title = NULL, content = NULL, placement = "top",
  ...)

use_bs_popover()

Arguments

tag
htmltools::tag, generally <button/> or <a/>, into which to embed the popover

title
character, title for the popover, generally text

content
character, content for the popover body, can be HTML

placement
character, placement of the popover with respect to tag

...
other named arguments, passed to bs_set_data()

Details

To activate the use of popovers in your page, you will need to call the use_bs_popover() function somewhere.

The verb embed is used to signify that you are embedding information into a tag. This implies that you can embed, at most, one "thing" into a particular tag. You should not, for example, expect to embed both a tooltip and a popover into a tag.

Value

htmltools::tag, modified copy of tag

See Also

bs_embed_tooltip, http://getbootstrap.com/javascript/#popovers

Examples

library("htmltools")

bs_button("A button") %>%
  bs_embed_popover(title = "I'm a popover", content = "Really!")

[Package bsplus version 0.1.1 Index]