Embed a tooltip into an html element

This is useful to add a few words of explanation to an element.

bs_embed_tooltip(tag, title = "", placement = c("right", "auto"), ...)

use_bs_tooltip()

Arguments

tag

htmltools::tag, button or link into which to embed a tooltip

title

character, title for the tooltip

placement

character, placement of the tooltip with respect to the tag

...

other named arguments, passed to bs_set_data()

Value

htmltools::tag, modified copy of tag

Details

To activate the use of tooltips in your page, you will need to call the use_bs_tooltip() function somewhere.

The verb "embed" is used to signify that we are embedding information into the tag. This implies that we can embed, at most, one "thing" into a particular tag. We should not, for example, embed both a tooltip and a popover into a tag.

See also

bs_embed_popover, http://getbootstrap.com/javascript/#tooltips

Examples

library("htmltools")
tags$button(type = "button", class = "btn btn-default", "I'm a button") %>%
  bs_embed_tooltip(title = "I'm a tooltip")

#> <button type="button" class="btn btn-default" title="I&#39;m a tooltip" data-toggle="tooltip" data-placement="right auto">I'm a button</button>