Tooltip

A tooltip can be a useful way to add a few words of explanation to a tag.

bs_embed_tooltip(tag, title = "", placement = "top", ...)

use_bs_tooltip()

Arguments

tag
htmltools::[tag][htmltools::tag], generally <button/> or <a/>, into which to embed the tooltip

title
character, title for the tooltip

placement
character, placement of the tooltip with respect to tag

...
other named arguments, passed to bs_set_data()

Value

htmltools::[tag][htmltools::tag], modified copy of tag

Details

To activate the use of tooltips in your page, you will need to call the use_bs_tooltip() function somewhere.

The verb embed is used to signify that you are embedding information into a tag. This implies that you can embed, at most, one "thing" into a particular tag. You should not, for example, expect to embed both a tooltip and a popover into a tag.

See also

Examples

library("htmltools")
bs_button("I'm a button") %>%
  bs_embed_tooltip(title = "I'm a tooltip")

#> <button class="btn btn-default" title="I&#39;m a tooltip" data-toggle="tooltip" data-placement="top">I'm a button</button>