Make a bootstrap modal window

You need to include the id as the `data-target` of a button or link.

Build a modal window

bs_modal(id, title, body, footer = bs_modal_closebutton(title = "Close"),
  size = c("medium", "large", "small"))

bs_modal(id, title, body, footer = bs_modal_closebutton(title = "Close"),
  size = c("medium", "large", "small"))

Arguments

id

character, unique identifier for the modal window

title

character (or HTML), to put in as the title of the modal window

body

character, content of the body, can be HTML fragment

footer

character, HTML fragment for footers

size

character, how big to make the window

...

elements to include in the body

id

character, unique id for modal window to create

title

character, title for the modal

id_modal

character, unique id modal window to attach

tag

character, HTML element to which to attach

Value

div containing the code for the modal window

HTML fragment

Examples

library("htmltools")

bs_modal(id = "modal", title = "I'm a modal", body = "Yes, I am.")

#> <div class="modal fade" id="modal" tabindex="-1" role="dialog" aria-labelledby="modal-title">
#>   <div class="modal-dialog" role="document">
#>     <div class="modal-content">
#>       <div class="modal-header">
#>         <button type="button" class="close" data-dismiss="modal" aria-label="Close">
#>           <span aria-hidden="true">&times;</span>
#>         </button>
#>         <h4 class="modal-title" id="modal-title">I'm a modal</h4>
#>       </div>
#>       <div class="modal-body">Yes, I am.</div>
#>       <div class="modal-footer">
#>         <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
#>       </div>
#>     </div>
#>   </div>
#> </div>

tags$button(type = "button", class = "btn btn-default", "Click for modal") %>%
  bs_attach_modal(id_modal = "modal")

#> <button type="button" class="btn btn-default" data-toggle="modal" data-target="#modal">Click for modal</button>