Create and attach a modal window

Modal windows are useful to make detailed explanations. If you want to compose body of a modal window using markdown, a helper function is provided: render_html_fragment.

bs_modal(id, title, body, footer = bs_modal_closebutton(title = "Close"),
  size = c("medium", "large", "small"))

bs_modal_closebutton(title = "Close")

bs_attach_modal(tag, id_modal)

Arguments

id

character, unique id for collapsible div

title

character, title for the modal window (or close-button)

body

character (HTML) or htmltools::tagList, content for the body of the modal window

footer

character (HTML) or htmltools::tagList, content for the footer of the modal window

size

character, size of the modal window

tag

htmltools::tag, button or link to which to attach the modal window

id_modal

character, unique id of modal window to attach

Value

bs_modal()

htmltools::tag, div

bs_attach_modal()

htmltools::tag, modified copy of tag

bs_modal_closebutton()

htmltools::tag, button

Details

Modal windows are typically attached to buttons or links. Thus, there are two parts to this system:

A modal window, created using bs_modal()

At least one button or link to which the id of the modal window is attached, using bs_attach_modal()

The verb "attach" is used to signify that we are attaching the id of our modal window to the tag in question (likely a button or a link). This implies that you can attach (the id of) a modal window to more than one button or link.

It is your responsibility to ensure that id is unique among html elements in your page. If you have non-unique id's, strange things may happen to your page.

Your code may be cleaner if you can import the content for the modal body from an external source. This is the motivation for the helper function, render_html_fragment, which is supplied with a path to a markdown (or even R markdown) file, then renders it into HTML, then returns the HTML. Accordingly, this can be supplied as the value of the body argument.

If you want to compose your own footer for the modal window, the function bs_modal_closebutton() can be useful. It takes a single argument, title, as the text to put in the button.

See also

render_html_fragment

Examples

library("htmltools")

bs_modal(id = "modal", title = "I'm a modal", body = "Yes, I am.")

#> <div class="modal fade" id="modal" tabindex="-1" role="dialog" aria-labelledby="modal-title">
#>   <div class="modal-dialog" role="document">
#>     <div class="modal-content">
#>       <div class="modal-header">
#>         <button type="button" class="close" data-dismiss="modal" aria-label="Close">
#>           <span aria-hidden="true">&times;</span>
#>         </button>
#>         <h4 class="modal-title" id="modal-title">I'm a modal</h4>
#>       </div>
#>       <div class="modal-body">Yes, I am.</div>
#>       <div class="modal-footer">
#>         <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
#>       </div>
#>     </div>
#>   </div>
#> </div>

tags$button(type = "button", class = "btn btn-default", "Click for modal") %>%
  bs_attach_modal(id_modal = "modal")

#> <button type="button" class="btn btn-default" data-toggle="modal" data-target="#modal">Click for modal</button>

bs_modal(
  id = "modal_large",
  title = "I'm a modal",
  size = "large",
  body = render_html_fragment(
    system.file("markdown", "modal.md", package = "bsplus")
  )
)

#> /Applications/RStudio.app/Contents/MacOS/pandoc/pandoc +RTS -K512m -RTS modal.utf8.md --to html --from markdown+autolink_bare_uris+ascii_identifiers+tex_math_single_backslash --output /var/folders/rg/9kvy_lsn43z90d439n02nfw80000gn/T//RtmppEmVaX/file81366cdea20b --smart --email-obfuscation none --self-contained --mathjax --standalone --section-divs --template /Library/Frameworks/R.framework/Versions/3.3/Resources/library/rmarkdown/rmd/fragment/default.html --no-highlight 

#> 
#> Output created: /var/folders/rg/9kvy_lsn43z90d439n02nfw80000gn/T//RtmppEmVaX/file81366cdea20b

#> <div aria-labelledby="modal_large-title" class="modal fade bs-example-modal-lg" id="modal_large" role="dialog" tabindex="-1">
#>   <div class="modal-dialog modal-lg" role="document">
#>     <div class="modal-content">
#>       <div class="modal-header">
#>         <button type="button" class="close" data-dismiss="modal" aria-label="Close">
#>           <span aria-hidden="true">&times;</span>
#>         </button>
#>         <h4 class="modal-title" id="modal_large-title">I'm a modal</h4>
#>       </div>
#>       <div class="modal-body"><div id="th-level-headers-work-well" class="section level4"> <h4>4th-level headers work well</h4> <p>We can do our regular Markdown stuff in modal windows.</p> <ul> <li><span class="math inline">\(F = ma\)</span></li> <li><span class="math inline">\(E = mc^2\)</span></li> </ul> </div></div>
#>       <div class="modal-footer">
#>         <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
#>       </div>
#>     </div>
#>   </div>
#> </div>

tags$button(type = "button", class = "btn btn-default", "Click for modal") %>%
  bs_attach_modal(id_modal = "modal_large")

#> <button type="button" class="btn btn-default" data-toggle="modal" data-target="#modal_large">Click for modal</button>