Renders and returns an HTML fragment

This is a wrapper around the rmarkdown::render function. The principal difference is that the function is designed to return an HTML fragment (rather than writing to a file). This function is useful to populate the content of a modal window.

render_html_fragment(input, output_format = rmarkdown::html_fragment(), ...)

Arguments

input

character, path to input file

output_format

rmarkdown output format, provided so you can specify arguments

...

other arguments passed to rmarkdown::render

Value

htmltools::tag

Examples

my_file <- system.file("markdown", "modal.md", package = "bsplus")
render_html_fragment(my_file)

#> /Applications/RStudio.app/Contents/MacOS/pandoc/pandoc +RTS -K512m -RTS modal.utf8.md --to html --from markdown+autolink_bare_uris+ascii_identifiers+tex_math_single_backslash --output /var/folders/rg/9kvy_lsn43z90d439n02nfw80000gn/T//Rtmp9qzeBm/file6e275cf78e13 --smart --email-obfuscation none --self-contained --mathjax --standalone --section-divs --template /Library/Frameworks/R.framework/Versions/3.3/Resources/library/rmarkdown/rmd/fragment/default.html --no-highlight 

#> 
#> Output created: /var/folders/rg/9kvy_lsn43z90d439n02nfw80000gn/T//Rtmp9qzeBm/file6e275cf78e13

#> <div id="th-level-headers-work-well" class="section level4"> <h4>4th-level headers work well</h4> <p>We can do our regular Markdown stuff in modal windows.</p> <ul> <li><span class="math inline">\(F = ma\)</span></li> <li><span class="math inline">\(E = mc^2\)</span></li> </ul> </div>