Renders and returns an HTML fragment

This is a wrapper around the rmarkdown::[render][rmarkdown::render] function. The principal difference is that the function is designed to return an HTML fragment (rather than writing to a file). This function is useful to populate the content of a modal window.

render_html_fragment(input, output_format = rmarkdown::html_fragment(), ...)

Arguments

input
character, path to input file

output_format
rmarkdown output format, provided so you can specify arguments

...
other arguments passed to rmarkdown::[render][rmarkdown::render]

Value

htmltools::[tag][htmltools::tag]

Details

This function is being deprecated in favor of shiny::[includeMarkdown][shiny::includeMarkdown]

Examples

if (FALSE) {
my_file <- system.file("markdown", "modal.md", package = "bsplus")
render_html_fragment(my_file)
}